youtube-analytics-api-setup/requirements.txt
# YouTube API Dependencies

# Google APIs
google-api-python-client>=2.95.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# YouTube transcript extraction
youtube-transcript-api>=0.6.1

# Environment management
python-dotenv>=1.0.0

# Data processing
pathlib>=1.0.1

# Optional but useful
requests>=2.31.0
